requirements.txt
asgiref==3.7.2
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.7
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
dulwich==0.21.6
fastjsonschema==2.18.1
filelock==3.12.2
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
more-itertools==10.1.0
msgpack==1.0.7
packaging==23.2
pexpect==4.8.0
pillow==10.2.0
pipenv==2023.11.15
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.3
sqlparse==0.4.4
tomlkit==0.12.1
trove-classifiers==2023.9.19
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.25.0
zipp==3.17.0
